[UPDATE] get_simulator method
