+++ src/mewpy/omics/integration/gimme.py
@@ -26,7 +26,6 @@ from copy import deepcopy
 from mewpy.solvers.solution import to_simulation_result
 from mewpy.solvers import solver_instance
 from mewpy.simulation import get_simulator
-from mewpy.simulation.simulation import Simulator
 from mewpy.cobra.util import convert_to_irreversible
 from .. import Preprocessing, ExpressionSet
 
@@ -59,15 +58,9 @@ def GIMME(model, expr, biomass=None, condition=0, cutoff=25, growth_frac=0.9,
            doi:10.1371/journal.pcbi.1000082
     """
     if not inline:
-        if isinstance(model, Simulator):
-            sim = model
-        else:
-            sim = get_simulator(model)
+        sim = get_simulator(model)
     else:
-        if isinstance(model, Simulator):
-            sim = deepcopy(model)
-        else:
-            sim = get_simulator(deepcopy(model))
+        sim = get_simulator(deepcopy(model))
         
     if isinstance(expr, ExpressionSet):
         pp = Preprocessing(sim, expr)
