+++ src/mewpy/problems/com.py
@@ -66,7 +66,7 @@ class CommunityKOProblem(AbstractKOProblem):
             warn("Model ids are not unique, repeated models will be discarded.")
 
         for model in models:
-            m = model if isinstance(model, Simulator) else get_simulator(model)
+            m = get_simulator(model)
             self.organisms[m.id] = deepcopy(m) if copy_models else m
 
         self.cmodel = CommunityModel(list(self.organisms.values()), flavor=self.flavor)
