+++ src/mewpy/solvers/solver.py
@@ -25,8 +25,7 @@ from enum import Enum
 from math import inf
 from reframed.core.cbmodel import CBModel
 from cobra.core.model import Model
-
-from ..simulation.simulation import Simulator
+from ..simulation import get_simulator
 
 
 class VarType(Enum):
@@ -172,21 +171,7 @@ class Solver(object):
             model
         """
 
-        if isinstance(model, (CBModel, Model)):
-            self.__build_problem_model(model)
-        elif isinstance(model, Simulator):
-            self.__build_problem_simulator(model)
-        else:
-            raise TypeError
-
-    def __build_problem_model(self, model):
-        """ Create a problem for metabolic models (REFRAMED or COBRApy)
-        Args:
-            model: A metabolic model
-        """
-        from ..simulation import get_simulator
-        sim = get_simulator(model)
-        self.__build_problem_simulator(sim)
+        self.__build_problem_simulator(get_simulator(model))
 
     def __build_problem_simulator(self, simulator):
         """Create a problem for simulators
