+++ src/mewpy/visualization/envelope.py
@@ -18,9 +18,7 @@
    Author: Vitor Pereira
 """
 import numpy as np
-
-from ..simulation import get_simulator
-from ..simulation.simulation import Simulator
+from mewpy.simulation import get_simulator
 
 
 def flux_envelope(model, r_x, r_y, steps=10, constraints=None, x_range=None, tolerance=0):
@@ -39,14 +37,11 @@ def flux_envelope(model, r_x, r_y, steps=10, constraints=None, x_range=None, tol
 
     """
 
-    if isinstance(model, Simulator):
-        simul = model
-    else:
-        try:
-            simul = get_simulator(model)
-        except Exception:
-            raise ValueError(
-                'The model should be an instance of model or simulator')
+    try:
+        simul = get_simulator(model)
+    except Exception:
+        raise ValueError(
+            'The model should be an instance of model or simulator')
 
     obj_frac = 0
     # if r_x in simul.get_objective():
@@ -105,14 +100,11 @@ def plot_flux_envelope(model, r_x, r_y, steps=10, substrate=None, constraints=No
     except ImportError:
         raise RuntimeError("Matplotlib is not installed.")
 
-    if isinstance(model, Simulator):
-        simul = model
-    else:
-        try:
-            simul = get_simulator(model)
-        except Exception:
-            raise ValueError(
-                'model should be an instance of model or simulator')
+    try:
+        simul = get_simulator(model)
+    except Exception:
+        raise ValueError(
+            'model should be an instance of model or simulator')
 
     offset = 0.03
